Add tests for current_user_backend.
