+++ tests/astronomer/flask_appbuilder/conftest.py
@@ -1,12 +1,16 @@
+import os
+import time
 import uuid
 
-import flask_appbuilder
-from jwcrypto import jwk
 import pytest
-from sqlalchemy import event
-
 from astronomer.flask_appbuilder import current_user_backend
 from astronomer.flask_appbuilder.security import AstroSecurityManagerMixin
+from jwcrypto import jwk, jwt
+from sqlalchemy import event
+
+import flask_appbuilder
+
+AUDIENCE = 'airflow.example.com'
 
 @pytest.fixture(scope='module')
 def app():
@@ -159,3 +163,46 @@ def role(appbuilder):
         return role
 
     return role_factory
+
+
+
+@pytest.fixture(scope='session')
+def allowed_audience():
+    return AUDIENCE
+
+
+@pytest.fixture
+def invalid_jwt():
+    ephemeral_key = jwk.JWK(generate='oct', size=64)
+    token = jwt.JWT(header={"alg": "HS256"}, claims={"info": "I'm a signed token"})
+    token.make_signed_token(ephemeral_key)
+    return token.serialize()
+
+
+@pytest.fixture
+def signed_jwt(jwt_signing_key):
+    def jwt_factory(claims):
+        token = jwt.JWT(header={"alg": "HS256"}, claims=claims)
+        token.make_signed_token(jwt_signing_key)
+        return token.serialize()
+    return jwt_factory
+
+
+@pytest.fixture
+def valid_claims(request):
+
+    return {
+        'email': 'airflower@domain.com',
+        'roles': ['Op'],
+        'sub': str(uuid.uuid4()),
+        'full_name': 'Lucy Airflower',
+        'aud': AUDIENCE,
+        'nbf': int(time.time()),
+        'exp': int(time.time()) + 60,
+    }
+
+
+@pytest.fixture
+def airflow_config(jwt_signing_cert, monkeypatch, allowed_audience):
+    monkeypatch.setitem(os.environ, 'AIRFLOW__ASTRONOMER__JWT_SIGNING_CERT', jwt_signing_cert)
+    monkeypatch.setitem(os.environ, 'AIRFLOW__ASTRONOMER__JWT_AUDIENCE', allowed_audience)
